+++ p2p_wallet/Scenes/Main/NewHome/HomeCoordinator.swift
@@ -417,10 +417,14 @@ final class HomeCoordinator: Coordinator<Void> {
     }
 
     private func openPendingTransactionDetails(transaction: PendingTransaction) -> AnyPublisher<Void, Never> {
-        coordinate(to: TransactionDetailCoordinator(
-            viewModel: TransactionDetailViewModel(pendingTransaction: transaction),
-            presentingViewController: navigationController
-        ))
+        // We need this delay to handle pop animation
+        Just(()).delay(for: 0.8, scheduler: RunLoop.main)
+            .flatMap { [unowned self] in
+            coordinate(to: TransactionDetailCoordinator(
+                viewModel: TransactionDetailViewModel(pendingTransaction: transaction),
+                presentingViewController: navigationController
+            ))
+        }
         .map { _ in () }
         .eraseToAnyPublisher()
     }
